requirements.txt
pandas==1.1.3
scipy==1.5.2
numpy==1.19.2
tqdm==4.50.2
librosa==0.8.0
requests==2.22.0
google-cloud-storage==1.36.0
